feat: add Github API interface for feign (#53)
