+++ src/main/java/scanner/dto/github/GitRepository.java
@@ -1,0 +1,11 @@
+package scanner.dto.github;
+
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class GitRepository {
+}
